updated all entities accordingly. have proper mapping and stuffs
though, the "columnDefinition" sets the SQL as i desire, but it doesn't create the default values that i want, pretty annoying.

+++ HealthyRecipe/src/main/java/com/FYP18/HealthyRecipe/Entity/BlogReviewRatingId.java
@@ -3,18 +3,20 @@ package com.FYP18.HealthyRecipe.Entity;
 import java.io.Serializable;
 
 import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-import jakarta.persistence.Id;
+import jakarta.persistence.Embeddable; 
 
 @Embeddable
 public class BlogReviewRatingId implements Serializable
 { 
     @Column(name= "UserID", nullable = false, updatable = false)
     public String UserID;
+    // this belongs to whoever user left review/rating
+    // usually registered user
  
+    
     @Column(name= "BlogID", nullable = false, updatable = false)
-    public Integer blogID;
-
+    public Long blogID;
+    // this belongs to which blog 
     @Override
     public String toString()
     {
